Arreglado arrf de publicidad

+++ analizador.py
@@ -1,10 +1,16 @@
 import sys, os, string
 import urllib.request
+import fnmatch
 import zipfile
 from subprocess import call
 from optparse import OptionParser
 from os import listdir
+from decimal import *
 import os
+import urllib.parse
+import urllib.request
+import glob
+
 
 carpetaPrincipal=os.path.dirname(os.path.abspath(__file__))
 carpetaAPK=carpetaPrincipal+"/Apk"
@@ -240,14 +246,158 @@ diccionarioPublicidad = {
 	
 }
 
+diccionarioEncriptacion={
+	"getInstance(\"AES\")":0,
+	"getInstance(\"AESWrap\")":0,
+	"getInstance(\"Blowfish\")":0,
+	"getInstance(\"CCM\")":0,
+	"getInstance(\"DES\")":0,
+	"getInstance(\"DESede\")":0,
+	"getInstance(\"CCM\")":0,
+	"getInstance(\"DESedeWrap\")":0,
+	"getInstance(\"ECIES\")":0,
+	"getInstance(\"GCM\")":0,
+	"getInstance(\"RC2\")":0,
+	"getInstance(\"RC4\")":0,
+	"getInstance(\"RC5\")":0,
+	"getInstance(\"RSA\")":0,
+
+
+}
+
 permisosBuscados = []
 
 listaPermisos = []
 listaPublicidad = []
+listaEncriptacion =[]
+
+data =[]
+
+diccionarioRiesgoPermisos={
+
+	"READ_SMS": 0.4428,
+	"RECEIVE_SMS": 0.42,
+	"SEND_SMS": 0.3961,
+	"WRITE_SMS": 0.1988,
+	"SET_ALARM": 0.1443,
+	"RECEIVE_WAP_PUSH": 0.1403,
+	"READ_PHONE_STATE": 0.1140,
+	"READ_EXTERNAL_STORAGE": 0.1044,
+	"RESTART_PACKAGES": 0.0804,
+	"SYSTEM_ALERT_WINDOW": 0.0711,
+	"RECEIVE_BOOT_COMPLETED": 0.0668,
+	"CHANGE_WIFI_STATE": 0.0630,
+	"WAKE_LOCK": 0.0611,
+	"DISABLE_KEYGUARD": 0.0562,
+	"ACCESS_NETWORK_STATE": 0.0553,
+	"WRITE_SETTINGS": 0.0551,
+	"READ_CONTACTS": 0.0535,
+	"RECEIVE_MMS": 0.0530,
+	"WRITE_EXTERNAL_STORAGE": 0.0506,
+	"EXPAND_STATUS_BAR": 0.0450,
+	"WRITE_CONTACTS": 0.0444,
+	"CHANGE_NETWORK_STATE": 0.0415,
+	"INTERNET": 0.0413,
+	"READ_HISTORY_BOOKMARKS": 0.0365,
+	"CHANGE_CONFIGURATION": 0.0346,
+	"PROCESS_OUTGOING_CALLS": 0.0344,
+	"GET_PACKAGE_SIZE": 0.0339,
+	"PERSISTENT_ACTIVITY": 0.0338,
+	"ACCESS_WIFI_STATE": 0.0334,
+	"READ_CALL_LOG": 0.0329,
+	"CAMERA": 0.0309,
+	"WRITE_HISTORY_BOOKMARKS":0.0287,
+	"CALL_PHONE": 0.0273,
+	"SET_WALLPAPER_HINTS": 0.0252,
+	"GET_ACCOUNTS": 0.0249,
+	"GET_TASKS": 0.0237,
+	"WRITE_CALL_LOG": 0.0232,
+	"ADD_SYSTEM_SERVICE": 0.0190,
+	"ACCESS_FINE_LOCATION": 0.0182,
+	"ACCESS_MOCK_LOCATION": 0.0168,
+
+	"SumatoriaRiesgos": 3.2764
+	
+}
+
+
+diccionarioRiesgoPublicidad = {
+
+	"com.google.ads": 6.427786595,
+	"com.flurry": 12.66634111,
+	"com.inmobi": 7.935539006,
+	"com.tapjoy": 7.447198144,
+	"com.mobclix": 6.427786595,
+	"com.chartboost": 3.967769503,
+	"com.adwhirl": 5.808204126,
+	"com.mopub": 9.186912465,
+	"com.greystripe": 4.49273593,
+	"com.jumptap": 6.9405445,
+	"com.google.analytics": 12.66634111,
+	"com.admob": -1,
+	"com.burstly": 8.99157612,
+	"com.sponsorpay": 7.447198144,
+	"com.cauly": 2.948357954,
+	"com.mobfox": 3.503845684,
+	"com.vpon": 9.504334025,
+	"com.appbrain": 2.948357954,
+	"net.daum": 10.89305335,
+	"com.admarvel": 10.05982176,
+	"com.applovin": 7.972164571,
+	"com.adfonic": 6.427786595,
+	"com.mdotm": 11.15858869,
+	"com.getjar": 7.911121963,
+	"com.nexage": 2.948357954,
+	"com.inneractive": 9.040410206,
+	"com.pontiflex": -1,
+	"com.zestadz": 1.26052985,
+	"com.madhouse": -1,
+	"com.smaato": 12.43132707,
+	"net.youmi": 11.88499573,
+	"de.madvertise": 2.328775485,
+	"cn.domob": 8.698571603,
+	"com.jirbo.adcolony": 2.804907826,
+	"com.revmob": 7.447198144,
+	"com.senddroid": 4.739958491,
+	"com.airpush": 7.972164571,
+	"com.tapit": 7.49603223,
+	"com.medialets": -1,
+	"mediba.ad": 3.885361983,
+	"com.papaya": 6.427786595,
+	"com.huntmads": -1,
+	"com.rhythmnewmedia": 5.213038701,
+	"com.tapfortap": 11.65303382,
+	"com.adknowledge": -1,
+	"net.metaps": 7.447198144,
+	"com.wiyun": -1,
+	"com.vdopia": 2.948357954,
+	"com.waps": -1,
+	"com.adwo": -1,
+	"com.mobosquare": -1,
+	"mobi.vserv": -1,
+	"com.wooboo": -1,
+	"com.everbadge": -1,
+	"com.mt.airad": 5.560981565,
+	"com.noqoush.adfalcon": 8.823708949,
+	"com.moolah": 1.26052985,
+	"com.kuguo": 7.959956049,
+	"buzzcity": -1,
+	"com.adsmogo": -1,
+	"com.sellaring": -1,
+	"com.startapp": 4.691124405,
+	"com.admoda": -1,
+	"com.mobpartner": 9.186912465,
+	"com.quclix": -1,
+	"com.ldevelop": -1,
+
+	"SumatoriaRiesgosPublicidad": 329.7094372
+	
+}
 
 
 
-def revisarManifest(folder,carpetaArchivosJava):
+def revisarManifest(folder,carpetaArchivosJava,MalWare):
+
 
 	f=open(folder+'/AndroidManifest.xml')
 	lines=f.readlines()
@@ -304,13 +454,15 @@ def revisarManifest(folder,carpetaArchivosJava):
 
 			permissionName = permissionName[0:n]
 
-		
-			if permissionName not in listaPermisos:
-				listaPermisos.append(permissionName)
+			if permissionName in diccionarioPermisos:
+				if permissionName not in listaPermisos:
+					listaPermisos.append(permissionName)
 
 
 	print('Package Name: '+packageName)
 	print("")
+
+	acum = 0
 	
 	for x in listaPermisos:
 
@@ -319,22 +471,27 @@ def revisarManifest(folder,carpetaArchivosJava):
 		
 		diccionarioPermisos[x] = 1;
 
-	if(MalWare):
+		if x in diccionarioRiesgoPermisos:
+			#print (x +'    '+str(diccionarioRiesgoPermisos.get(x)))
+			acum = acum + diccionarioRiesgoPermisos.get(x)
 
-		os.chdir("output_Permisos-Malware")
-		archivo=open('outputPermisos-Malware.arff' ,'a')
+	#print('\n' + str(acum))
 
-	else:
+	riesgoPermisos = (acum * 100)/diccionarioRiesgoPermisos.get("SumatoriaRiesgos")
 	
-		os.chdir("output_Permisos")		
+	#print(riesgoPermisos)
 
-		archivo=open('outputPermisos.arff' ,'a')
+	os.chdir(carpetaPrincipal)
+	
 
+	archivo=open('ResultadosApp.arff' ,'a')
 	
 
 	archivo.write(packageName+',')
 	archivo.write(Version + ',')
 
+
+	contador = 1
 
 	for n in diccionarioPermisos:
 
@@ -347,6 +504,8 @@ def revisarManifest(folder,carpetaArchivosJava):
 		else:
 			archivo.write('true,')
 
+		contador=contador+1
+
 	for x in listaPermisos:
 
 		ubicacion = listaPermisos.index(x)
@@ -357,14 +516,52 @@ def revisarManifest(folder,carpetaArchivosJava):
 	listaPermisos.clear()
 			
 
-	archivo.write('\n')
-
 	archivo.close()	
 
 
 	os.chdir(carpetaPrincipal)
 
-	revisarJava(carpetaArchivosJava,packageName,Version)
+	intento=1
+
+	revisarJava(carpetaArchivosJava,packageName,Version,MalWare,riesgoPermisos,intento)
+
+	
+def escribirPrimeraParte():
+
+	archivo=open('ResultadosApp.arff' ,'w')
+
+	archivo.write('\n')
+	archivo.write('@Relation AppResults')
+	archivo.write('\n')
+	archivo.write('\n')
+	archivo.write('@ATTRIBUTE packageName string\n')
+	archivo.write('@ATTRIBUTE Version string\n')
+
+	for n in diccionarioPermisos:
+		archivo.write('@ATTRIBUTE '+ n + ' {true,false}\n')
+
+	for n in diccionarioPublicidad:
+		archivo.write('@ATTRIBUTE '+ n + ' {true,false}\n')
+
+	for n in diccionarioEncriptacion:
+	
+		k=13
+
+		crypto=n[13:]
+
+		n = crypto.find('"')
+
+		crypto = crypto[0:n]
+
+		archivo.write('@ATTRIBUTE '+ crypto + ' {true,false}\n')
+
+	archivo.write('@ATTRIBUTE class {malware,normal}\n')
+	archivo.write('\n')
+	archivo.write('\n')
+	archivo.write('@data')
+	archivo.write('\n')
+	archivo.write('\n')
+	
 
    	
 
@@ -375,101 +572,12 @@ def ficherosEnDir(ruta):
         for fichero in ficheros:
             listado.append(os.path.join(raiz,fichero))
     return listado
-	
-
-def escribirPrimeraParte():
 
 
-	os.chdir("output_Permisos")		
 
-	archivo=open('outputPermisos.arff' ,'w')
+def revisarJava(folder,paquete,Version,MalWare,riesgoPermisos,intento):
 
-	archivo.write('\n')
-	archivo.write('@Relation Permissions')
-	archivo.write('\n')
-	archivo.write('\n')
-	archivo.write('@ATTRIBUTE packageName string\n')
-	archivo.write('@ATTRIBUTE Version string\n')
-	
-
-	for n in diccionarioPermisos:
-		archivo.write('@ATTRIBUTE '+ n + ' {true,false}\n')
-
-	archivo.write('\n')
-	archivo.write('\n')
-	archivo.write('@data')
-	archivo.write('\n')
-	archivo.write('\n')
-
-	os.chdir(carpetaPrincipal)
-
-	os.chdir("output_Permisos-Malware")		
-
-	archivo=open('outputPermisos-Malware.arff' ,'w')
-
-	archivo.write('\n')
-	archivo.write('@Relation Permissions')
-	archivo.write('\n')
-	archivo.write('\n')
-	archivo.write('@ATTRIBUTE packageName string\n')
-	archivo.write('@ATTRIBUTE Version string\n')
-	
-
-	for n in diccionarioPermisos:
-		archivo.write('@ATTRIBUTE '+ n + ' {true,false}\n')
-
-	archivo.write('\n')
-	archivo.write('\n')
-	archivo.write('@data')
-	archivo.write('\n')
-	archivo.write('\n')
-
-	os.chdir(carpetaPrincipal)
-
-	os.chdir("output_Publicidad")			
-	archivo2=open('outputPublicidad.arff' ,'w')
-	
-	archivo2.write('\n')
-	archivo2.write('@Relation Publicity')
-	archivo2.write('\n')
-	archivo2.write('\n')
-
-	archivo2.write('@ATTRIBUTE packageName string\n')
-	archivo2.write('@ATTRIBUTE Version string\n')
-
-	for n in diccionarioPublicidad:
-		archivo2.write('@ATTRIBUTE '+ n + ' {true,false}\n')
-
-	archivo2.write('\n')
-	archivo2.write('\n')
-	archivo2.write('@data')
-	archivo2.write('\n')
-
-	os.chdir(carpetaPrincipal)
-
-	os.chdir("output_Publicidad-Malware")			
-	archivo2=open('outputPublicidad-Malware.arff' ,'w')
-	
-	archivo2.write('\n')
-	archivo2.write('@Relation Publicity')
-	archivo2.write('\n')
-	archivo2.write('\n')
-
-	archivo2.write('@ATTRIBUTE packageName string\n')
-	archivo2.write('@ATTRIBUTE Version string\n')
-
-	for n in diccionarioPublicidad:
-		archivo2.write('@ATTRIBUTE '+ n + ' {true,false}\n')
-
-	archivo2.write('\n')
-	archivo2.write('\n')
-	archivo2.write('@data')
-	archivo2.write('\n')
-
-	os.chdir(carpetaPrincipal)
-
-
-def revisarJava(folder,paquete,Version):
+	intento = intento +1
 
 	listaArchivos=ficherosEnDir(folder)
 
@@ -483,38 +591,54 @@ def revisarJava(folder,paquete,Version):
 				if z.find(y) != -1:
 					if y not in listaPublicidad:
 						listaPublicidad.append(y)
-					break;
+		
+		for o in diccionarioEncriptacion:
+			for q in lines:
+				if q.find(o) !=-1:
+					if o not in listaEncriptacion:
+						listaEncriptacion.append(o)
+
 
 		javaFile.close()
 
+	riesgoPublicidad=0
+
+	riesgoEncriptacion=0
+
 	for g in listaPublicidad:
+
 		ubicacion = listaPublicidad.index(g)
+
+		if diccionarioRiesgoPublicidad.get(g)==-1:
+			riesgoPublicidad=-1
+			break
+
+		else:
+
+			riesgoPublicidad=riesgoPublicidad+diccionarioRiesgoPublicidad.get(g)
+
+			riesgoPublicidad = (riesgoPublicidad * 100)/diccionarioRiesgoPublicidad.get("SumatoriaRiesgosPublicidad")
 
 		diccionarioPublicidad[g] = 1;
 
-
-
-	#for q in diccionarioPublicidad:
-		#print(q + diccionarioPublicidad.get(q)S)
-
 	
 
-	#print(diccionarioPublicidad)
-		#print(x)
+	
+	if len(listaEncriptacion)!=0:
 
-	if(MalWare):
-		os.chdir("output_Publicidad-Malware")			
-		archivo=open('outputPublicidad-Malware.arff' ,'a')
+		riesgoEncriptacion=1
+	
 
-	else:
-		os.chdir("output_Publicidad")			
-		archivo=open('outputPublicidad.arff' ,'a')
+	for h in listaEncriptacion:
+		location = listaEncriptacion.index(h)
 
+		diccionarioEncriptacion[h] = 1;	
 
-	archivo.write('\n')
+	archivo=open('ResultadosApp.arff' ,'a')
 
-	archivo.write(paquete+',')
-	archivo.write(Version + ',')
+	#for f in diccionarioPublicidad:
+		#print(f + "    "+ str(diccionarioPublicidad.get(f)))
+
 
 	for n in diccionarioPublicidad:
 
@@ -523,23 +647,79 @@ def revisarJava(folder,paquete,Version):
 		if(verificador==0):
 			archivo.write('false,')
 		
+		if(verificador==1):			
+			archivo.write('true,')
+			
+		diccionarioPublicidad[n] = 0;
+
+	listaPublicidad.clear()
+
+
+	for n in diccionarioEncriptacion:
+
+		verificador = diccionarioEncriptacion.get(n)
+
+		if(verificador==0):
+			archivo.write('false,')
+		
 		else:
 			archivo.write('true,')
 
+	for x in listaEncriptacion:
+
+		ubicacion = listaEncriptacion.index(x)
+
+		diccionarioEncriptacion[x] = 0;
+
+	listaEncriptacion.clear()
+
+	if MalWare:
+
+		archivo.write("malware")
+
+	else:
+		archivo.write("normal")
+
+	archivo.write('\n')
+
 	archivo.close()
 
-	os.chdir(carpetaPrincipal)
+	if riesgoPublicidad == -1:
+		riesgoTotal = riesgoPermisos
 
+	if riesgoPublicidad!=0:
+		riesgoTotal = (riesgoPermisos + riesgoPublicidad)/2
+
+	else:
+		riesgoTotal=riesgoPermisos
+
+	try:
+		url = 'http://adriangoicovic.16mb.com/database/post_all_apps.php'
+
+		values = {'nombre': paquete,
+		  		  'riesgoPermisos': riesgoPermisos,
+		 		  'riesgoEncriptacion': riesgoEncriptacion,
+				  'riesgoPublicidad': riesgoPublicidad,
+			 	  'riesgoTotal': riesgoTotal
+			 	  }
+
+		data = urllib.parse.urlencode(values)
+		data = data.encode('utf-8') 
+		req = urllib.request.Request(url, data)
+		response = urllib.request.urlopen(req)
+		the_page = response.read()
+		
+	except:
+		revisarJava(folder,paquete,Version,MalWare,riesgoPermisos,intento)
+
+	
 
 
 def validarCarpetas():
 
 	output = carpetaPrincipal+"/output"
 	output_Malware = carpetaPrincipal+"/output-Malware"
-	output_Permisos = carpetaPrincipal+"/output_Permisos"
-	output_Permisos_Malware = carpetaPrincipal+"/output_Permisos-Malware"
-	output_Publicidad = carpetaPrincipal+"/output_Publicidad"
-	output_Publicidad_Malware = carpetaPrincipal+"/output_Publicidad-Malware"
+		
 	
 	try:
 	    os.stat(output)
@@ -550,28 +730,36 @@ def validarCarpetas():
 	    os.stat(output_Malware)
 	except:
 	    os.mkdir(output_Malware) 
-
-	try:
-	    os.stat(output_Permisos)
-	except:
-	    os.mkdir(output_Permisos) 
-
-	try:
-	    os.stat(output_Permisos_Malware)
-	except:
-	    os.mkdir(output_Permisos_Malware) 
-
-	try:
-	    os.stat(output_Publicidad)
-	except:
-	    os.mkdir(output_Publicidad) 
-
-	try:
-	    os.stat(output_Publicidad_Malware)
-	except:
-	    os.mkdir(output_Publicidad_Malware) 
-
 	
+def ARFFvalido():
+
+	output = carpetaPrincipal+"/ResultadosApp.arff"
+
+	try:
+		os.stat(output)
+
+	except:
+
+		escribirPrimeraParte()
+
+def validarExistencia(nombreArchivo):
+
+	output = carpetaPrincipal+"/output/"+nombreArchivo
+
+	existe = True
+
+	try:
+		os.stat(output)
+
+		return existe
+
+	except:
+
+		existe=False
+
+		return existe
+
+
 
 
 def main():
@@ -580,14 +768,29 @@ def main():
 
 	validarCarpetas()
 
-	MalWare = False
+	ARFFvalido()
 
-	escribirPrimeraParte()
+	MalWare = False
 
 	ficheros = os.listdir(carpetaAPK)
 
 	ficherosMalware = os.listdir(carpetaAPKmalware)
 
+	matches = []
+	for root, dirnames, filenames in os.walk(carpetaAPK):
+	    for filename in fnmatch.filter(filenames, '*.apk'):
+	        matches.append(os.path.join(root, filename))
+	        #print(filename)
+
+	matchesMalware = []
+
+	for root, dirnames, filenames in os.walk(carpetaAPKmalware):
+		for filename in fnmatch.filter(filenames,'*.apk'):
+			matchesMalware.append(os.path.join(root, filename))
+	
+
+	#for h in diccionarioRiesgoPublicidad:
+		#print(h +"   "+str(diccionarioRiesgoPublicidad.get(h)))
 	
 
 	if not ficheros:
@@ -596,41 +799,47 @@ def main():
 
 	else:
 
-		for x in ficheros:
-			
-			nombreArchivo=os.path.splitext(x)[0]
-			
-			#python3 apk2java.py d /home/adrian/Escritorio/APK/com.whatsapp.apk -o hola --java
-			call('python3 decompilador.py d '+carpetaAPK +'/'+ x+' -o output/ ',shell=True)
-			call('rm -r -f output/'+nombreArchivo+'-new.apk',shell=True)
-			call('rm -r -f output/'+nombreArchivo+'.jar',shell=True)
+		for x in matches:
 
-			print ("*********************************************")
-			print ("**              Analizando                 **")
-			print ("*********************************************")
+				
+			nombreArchivo=os.path.splitext(os.path.basename(x))[0]
+
+			existe = validarExistencia(nombreArchivo)
+
+			if existe==False:
 			
-			carpetaArchivosJava=carpetaPrincipal+'/output/'+nombreArchivo+'/src'
+				#python3 apk2java.py d /home/adrian/Escritorio/APK/com.whatsapp.apk -o hola --java
+				call('python3 decompilador.py d '+x+' -o output/ ',shell=True)
+				call('rm -r -f output/'+nombreArchivo+'-new.apk',shell=True)
+				call('rm -r -f output/'+nombreArchivo+'.jar',shell=True)
 
-			revisarManifest(carpetaPrincipal+'/output/'+nombreArchivo,carpetaArchivosJava)
+				print ("*********************************************")
+				print ("**              Analizando                 **")
+				print ("*********************************************")
+			
+				carpetaArchivosJava=carpetaPrincipal+'/output/'+nombreArchivo+'/src'
 
-			print('')
-			print('Listo')
-			print('')
+				revisarManifest(carpetaPrincipal+'/output/'+nombreArchivo,carpetaArchivosJava,MalWare)
+
+				print('')
+				print('Listo')
+				print('')
 
 	if not ficherosMalware:
 		print("No hay MalWare")
 
+	
 	else:
 
 		MalWare = True;
 
-		for y in ficherosMalware:
+		for y in matchesMalware:
 
 
-			nombreArchivo=os.path.splitext(y)[0]
+			nombreArchivo=os.path.splitext(os.path.basename(y))[0]
 
 
-			call('python3 decompilador.py d '+carpetaAPKmalware +'/'+ y+' -o output-Malware',shell=True)
+			call('python3 decompilador.py d '+y+' -o output-Malware',shell=True)
 			call('rm -r -f output-Malware/'+nombreArchivo+'-new.apk',shell=True)
 			call('rm -r -f output-Malware/'+nombreArchivo+'.jar',shell=True)
 
@@ -640,12 +849,14 @@ def main():
 			
 			carpetaArchivosJava=carpetaPrincipal+'/output-Malware/'+nombreArchivo+'/src'
 
-			revisarManifest(carpetaPrincipal+'/output-Malware/'+nombreArchivo,carpetaArchivosJava)
+			revisarManifest(carpetaPrincipal+'/output-Malware/'+nombreArchivo,carpetaArchivosJava,MalWare)
+
+			
 
 			print('')
 			print('Listo')
 			print('')
-
+	
 
 if __name__=="__main__":
   main()
